♻️ Refactored Customer into its folder
